1. rename kratos init to kratos new 2. add kratos tool doc
