+++ README.md
@@ -8,7 +8,7 @@ Kratos是[bilibili](https://www.bilibili.com)开源的一套Go微服务框架，
 * [log](doc/wiki-cn/logger.md)：基于[zap](https://github.com/uber-go/zap)的field方式实现的高性能log库，集成了我们提供的[log-agent敬请期待]()日志收集方案。
 * [database](doc/wiki-cn/database.md)：集成MySQL&HBase&TiDB的SDK，其中TiDB使用服务发现方案。
 * [cache](doc/wiki-cn/cache.md)：集成memcache&redis的SDK，注意无redis-cluster实现，推荐使用代理模式[overlord](https://github.com/bilibili/overlord)。
-* [kratos tool](doc/wiki-cn/kratos-tool.md)：kratos相关工具量，包括项目快速生成、pb文件代码生成、swagger文档生成等。
+* [kratos工具](doc/wiki-cn/kratos-tool.md)：kratos相关工具量，包括项目快速生成、pb文件代码生成、swagger文档生成等。
 
 我们致力于提供完整的微服务研发体验，整合相关框架及工具后，微服务治理相关部分可对整体业务开发周期无感，从而更加聚焦于业务交付。对每位开发者而言，整套Kratos框架也是不错的学习仓库，可以了解和参考到[bilibili](https://www.bilibili.com)在微服务方面的技术积累和经验。
 
@@ -16,10 +16,11 @@ Kratos是[bilibili](https://www.bilibili.com)开源的一套Go微服务框架，
 
 ```shell
 go get -u github.com/bilibili/kratos/tool/kratos
-kratos init
+cd $GOPATH/src
+kratos new kratos-demo
 ```
 
-`kratos init`会快速生成基于kratos库的脚手架代码，如生成[kratos-demo](https://github.com/bilibili/kratos-demo)
+`kratos new kratos-demo`会快速生成基于kratos库的脚手架代码，如生成[kratos-demo](https://github.com/bilibili/kratos-demo)
 
 ```shell
 cd kratos-demo/cmd
@@ -29,9 +30,10 @@ go build
 
 打开浏览器访问：[http://localhost:8000/kratos-demo/start](http://localhost:8000/kratos-demo/start)，你会看到输出了`Golang 大法好 ！！！`
 
-[快速开始](doc/wiki-cn/quickstart.md)
+[快速开始](doc/wiki-cn/quickstart.md)  
+[kratos工具](doc/wiki-cn/kratos-tool.md)
 
-# Document
+# 文档目录
 
 [简体中文](doc/wiki-cn/summary.md)
