+++ tool/kratos/template.go
@@ -433,7 +433,7 @@ replace (
 	cloud.google.com/go => github.com/googleapis/google-cloud-go v0.26.0
 	golang.org/x/crypto => github.com/golang/crypto v0.0.0-20190123085648-057139ce5d2b
 	golang.org/x/lint => github.com/golang/lint v0.0.0-20181026193005-c67002cb31c3
-	golang.org/x/net => github.com/golang/net v0.0.0-20190311183353-d8887717615a
+	golang.org/x/net => github.com/golang/net v0.0.0-20190420063019-afa5a82059c6
 	golang.org/x/oauth2 => github.com/golang/oauth2 v0.0.0-20180821212333-d2e6202438be
 	golang.org/x/sync => github.com/golang/sync v0.0.0-20181108010431-42b317875d0f
 	golang.org/x/sys => github.com/golang/sys v0.0.0-20180905080454-ebe1bf3edb33
